% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorovForward.R
\name{feynman_kac}
\alias{feynman_kac}
\title{Solve a Feynman-Kac PDE under a Ito diffusion}
\usage{
feynman_kac(
  region,
  maturity,
  dynamics,
  problem,
  variational = FALSE,
  resolution = c(100, 100)
)
}
\arguments{
\item{region}{vector defining localized space region, an interval}

\item{maturity}{the time horizon}

\item{dynamics}{list of named coefficient functions defining the SDE, see details}

\item{problem}{list of functions defining the conditional expectation, see details}

\item{variational}{boolean for optional stopping problems}

\item{resolution}{the time and space resolution, as a vector, i.e. the number of sub-intervals in the discretization scheme}
}
\value{
a list containing the time-grid, space-grid, and solution matrix grid.
}
\description{
{Solves the Kolmogorov-forward PDE for a Feynman-Kac problem with arbitrary
discounting rate, running cost, terminal cost, and infinitesimal drift and diffusion coefficient functions.}
}
\details{
{The argument \code{dynamics} must be a named list containing functions of
time and space,\code{(t, x)}, where:
\itemize{
\item \code{drift} is the infinitesimal drift function,
\item \code{diffusion} is the infinitesimal volatility function}
and the argument \code{problem} must be a named list containing:
\itemize{
\item \code{discount} the discounting rate, a function of \code{(t, x)},
\item \code{running_cost} the running cost, a function of \code{(t, x)},
\item \code{terminal_cost}, the terminal cost, a function of \code{x}}
}
}
