% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorovForward.R
\name{gridFunction}
\alias{gridFunction}
\title{Discretize function over space-time grid}
\usage{
gridFunction(f, x, t = NULL)
}
\arguments{
\item{f}{function to evaluate on grid}

\item{x}{space variable}

\item{t}{time variable, can be null to discretize a function over just space.}
}
\value{
matrix
}
\description{
{Discretizes a function of space-time over a grid.}
}
\details{
{Simple double for-loop for space-time functions and one for-loop for functions of a single-variable}
}
