% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{indicator}
\alias{indicator}
\title{The indicator function of an event}
\usage{
indicator(bool)
}
\arguments{
\item{bool}{event or logical statement to indicate}
}
\value{
numeric
}
\description{
{Returns zero if the event does not occur and 1 if it does.}
}
