% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdes-package.R
\docType{package}
\name{pdes-package}
\alias{pdes}
\alias{pdes-package}
\title{pdes: Numerical solvers for parabolic PDEs, PIDEs and HJB-PDEs}
\description{
Solvers via Monte-Carlo, finite difference, and analytic solutions
    for parabolic PDE/PIDEs arising from Feynman-Kac or Fokker-Planck problems.
    Also provided are solvers for HJB-PDEs arising from stochastic control problems.
}
\keyword{internal}
